Allocator improvements

+++ nd4j-backends/nd4j-api-parent/nd4j-api/src/main/java/org/nd4j/linalg/api/ops/executioner/OpExecutionerUtil.java
@@ -25,7 +25,7 @@ public class OpExecutionerUtil {
 
         //Full buffer -> implies all elements are contiguous (and match)
         int l1 = x.length();
-        int dl1 = x.data().length();
+        long dl1 = x.data().length();
         if(l1 == dl1)
             return true;
 
@@ -45,9 +45,9 @@ public class OpExecutionerUtil {
         //Full buffer + matching strides -> implies all elements are contiguous (and match)
         //Need strides to match, otherwise elements in buffer won't line up (i.e., c vs. f order arrays)
         int l1 = x.length();
-        int dl1 = x.data().length();
+        long dl1 = x.data().length();
         int l2 = y.length();
-        int dl2 = y.data().length();
+        long dl2 = y.data().length();
         int[] strides1 = x.stride();
         int[] strides2 = y.stride();
         boolean equalStrides = Arrays.equals(strides1, strides2);
@@ -73,11 +73,11 @@ public class OpExecutionerUtil {
             return false;
         //Full buffer + matching strides -> implies all elements are contiguous (and match)
         int l1 = x.length();
-        int dl1 = x.data().length();
+        long dl1 = x.data().length();
         int l2 = y.length();
-        int dl2 = y.data().length();
+        long dl2 = y.data().length();
         int l3 = z.length();
-        int dl3 = z.data().length();
+        long dl3 = z.data().length();
         int[] strides1 = x.stride();
         int[] strides2 = y.stride();
         int[] strides3 = z.stride();
